Updated to handle BOOL values when they register with the NSMethodSignature method getArgumentTypeAtIndex:
I found that my returnTypeForKey:

    // bool
    BOOL aBOOL = YES;
    newType = [OBJCReturnTypes returnTypeForKey:@"bool"];

makes newType as "B", which processes just fine. However running the above

getArgumentTypeAtIndex:

where the argument type is a BOOL, returns "c" (meaning a char). A YES value was then showing up as "\x01" and a NO value showed up as "\0", neither of which translated properly when converting the value to a string. This is now supported by flipped the char value into an int, comparing against it and setting the values as literal chars of "1" and "0", which flips into a string very nicely. 0_0

Also updated the tests to check for this.

Next I'll set this up as a proper library module.