So in real life application the pointers are gone. Should've remembered that.
In any case, you now set the pointer data as a strong NSData object and the data type as a strong NSString.

This is the easiest usage, but there are a few optional methods I added in there if you don't like this one. Also in there is a new method to return the data type as an NSString instead of a char* (it's just a wrapper).

    // char *
    NSError *err;
    NSString *newType = [FiR3FiOBJCReturnTypes returnTypeAsStringForKey:@"char *"];
    NSString *stringResult = [parser parsePrimitiveWithData:[NSData dataWithBytes:aPtr length:strlen(aPtr)]
                                                    andType:newType
                                                      error:&err];